fixed exception when trying to delete non-existant file

+++ lib/util/localdata.dart
@@ -17,7 +17,9 @@ class LocalData {
 
   static Future<File> writeUserdata(String data) async {
     final file = await _localFile;
-    await file.delete();
+    if (await fileExists()) {
+      await file.delete();
+    }
     return file.writeAsString(data);
   }
